requirements.txt
atomicwrites==1.4.1
langchain==0.3.27
langchain-community==0.3.16
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-google-genai==2.0.0
langchain-tavily==0.2.11
langgraph==0.3.34
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.74
langsmith==0.4.29
pandas==2.3.2
pandocfilters==1.5.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
streamlit==1.37.1